* Added News Sentiment * Refactored Home Screen

+++ lib/entities/company_news_sentiment.dart
@@ -64,8 +64,8 @@ class Buzz {
 }
 
 class Sentiment {
-  int bearishPercent;
-  int bullishPercent;
+  double bearishPercent;
+  double bullishPercent;
 
   Sentiment({this.bearishPercent, this.bullishPercent});
